transcript second to float

+++ post.py
@@ -30,7 +30,7 @@ class Scores(BaseModel):
 
 
 class Transcript(BaseModel):
-    second: int
+    second: float
     text: str
